updateCollation: Drop 'order by cl_from' from query
We are removing the 'order by cl_from' from the select query in
updateCollation.php since it breaks the query in postgresql. The query
is completly fine without the order by since the order in which the
collations are handled does not matter.

Partial backport of I63a620a89c4ac8a85c263a4b7862af40e080c68e

Change-Id: I2eeae5e8badabf5dd1072284297a2e681d63fedc