selenium: Enable retry metrics per test in Prometheus for core
With the retry flag, retries will happen within the same runner
so that we can get statistics about it per test. If that fails
we will still do a retry on spec level (meaning we get metrics
on project level).

Lets test this for now in core and see that we can get valuable
information out of it.

Bug: T399677
Change-Id: I0683db96a707e6d25ba23eacb5f79562486a0263