docs: Mention deprecated Skin::appendSpecialPagesLinkIfAbsent in release notes
With T333211, we added a deprecation warning to a temporary
migration-easing function to alert third-party wikis to migrate the f ff
location of the Special:SpecialPages link.

This small commit simply puts the name of that function in the relevant
changelog entry to allow server admins to more easily find the reason
this deprecation warning is emitted.

Bug: T386761
Change-Id: I0affd0324e7610f51e68a06527634c4f6b50ab8d