filebackend: Remove consistency check for multi-backend
This is making uploads considerably slower: By one measurment, around
46% of the upload time is being spent on this function (410ms) which in turn
increases the chance of one of the requests failing and leaving the
system in an inconsistent state defying the whole point of this function
existing in the first place.

See:
https://performance.wikimedia.org/excimer/profile/03c4168ac24eb7eb

On top of this, MW filebackend is not responsible for ensuring
consistency of swift backends and is another classic case of mw doing
the job of the infrastructure. Swift infra has its own system of making
sure the backends are consistent and this function by drastically
increasing the upload time is actually making things worse for that
system.

To make an analogy, this is like mediawiki doing the work of db
replication to a remote datacenter.

Eventually, all of FileBackendMultiWrite needs to fully go away in favor
of a S3-compliant middleware but that's a next-year me problem.

Bug: T328872
Change-Id: I2a8d984bd45ad52094a0ef74a45afcff5819b1e6
(cherry picked from commit f7ea531b2f3d5c82edecf8881bd3264cddd9fac1)