Use a JWT cookie in CookieSessionProvider
Emit a JWT cookie (under the name 'sessionJwt') based on
SessionManager::getJwtData() added in I9240cac29f0209778.
The behavior is behind the $wgSessionCookieJwt feature flag.

The cookie uses an unprefixed name, as the goal is to make it
easy for edge infrastructure to partially verify sessions, and
the usual wiki-id prefix is hard for external infrastructure
to handle accurately.

For now, the cookie is added on top of the normal cookies, is
not required to be present in a valid session, and is not a valid
source of authentication on its own, but its consistency with the
session cookies is validated.
In the future the JWT cookie might replace the normal cookies
entirely, but that requires some caution as many tools have
hardcoded the names or format of MediaWiki cookies

The code is written in such a way that subclasses can easily
modify the characteristics of the cookie. The assumption is
that multiple wikis will share the same cookie for non-core
providers (in practice, CentralAuth). The JWT is kept free
of anything wiki- or session-specific to enable this.

The patch must be merged at the same time as the corresponding
CentralAuth change (Ic3a40b729c5899e7) due to the constructor
argument changes.

Bug: T399200
Change-Id: Ida66d491372b92390baaf4cc8de6e1696c9d6b96