Do not use UserNameUtils in SessionBackend::save()
Sessions are saved when they are deregistered, which happens at
unpredictable times, so avoid fetching a service that might not
be available yet.

The check was added in I12ff84db61 to differentiate between
anonymous users and users who are registered on another wiki
but have no local account (ie. between `new User()` style user
objects and `User::newFromName(...)` style user objects) and
an IP regex check, which can be done without a service, is good
enough for that.

The behavior is probably not identical to the old one, e.g. when
different wikis share session cookies but have different username
validity rules, but the differences are limited to fringe
scenarios, and the new behavior makes more sense - SessionBackend
should not handle invalid usernames differently when the session
is verified against a saved object vs. when it isn't.

Bug: T400113
Change-Id: Ib615e00bdf5daa060979ad747ed00d604b307d96