Minimize serialization footprint of MessageValue parameters
At the moment all parameters are serialized as individual
ScalarParam instances. This adds quite a lot of overhead, even if
the vast majority of parameters are typically trivial strings and
ints.

Before:
 "params":{"0":{"text":"hello",
 "_type_":"Wikimedia\\\\Message\\\\ScalarParam"},"_type_":"array"}

After:
 "params":["hello"]

Compatibility: The de-serialization just calls the constructor. The
constructor was always able to accept scalar values. The resulting
MessageValue instances are all the same as before.

Performance: Should not be affected at all. In both cases the
(almost) same code is called to re-construct the original
MessageValue with all original ScalarParam instances.

Depends-On: I9c683c109eabfe8da50f3555f5b7e9dfee5f002c
Change-Id: Ia40d13c35ac6dffa9f557c3784c37af0e42fcacc