DEVELOPERS: add note about AMD64 runtime and remove ancient macOS advice
Add a note in the requirements section that the container images
currently provided are only for an AMD64 runtime. These are known to
work on ARM64 hardware with macOS' Rosetta emulation, but are unlikely
to work on any 32-bit processor.

Remove "Mac users" section added when folks were having issues with
Apple M1 processors and early Docker Desktop Rosetta integration.

Bug: T355973
Change-Id: I5e7ba3bd1f9fe8697a543bfe9ab6189079d58ba5