Fix performer logging in AuthManager
We need to set the performer to the session user before we
overwrite the session user.

The only hook handler known to CodeSearch which is affected by this is
WikimediaEventsHooks::onAuthManagerLoginAuthenticateAudit() which uses
it to set the performer field on the mediawiki.accountcreation.login
event stream.

Follows-up: 487c0da87a7215d3a49a8c432109205f8789247b
Change-Id: Idf52fb80a91eef4ccb46179e4ab06237af8a5648