Remove vestigial $wgInitialSessionId and related code
$wgInitialSessionId was only used in a single check on
SpecialUserLogin/SpecialCreateAccount that is unreachable,
because it's impossible to reach the success screen
of those pages while not having cookies enabled.

Its intended purpose was to display a nice error message when
the user's browser has cookies disabled. This is already handled
elsewhere; I added a TODO note there to use the message specific
to the session provider, but I'm not changing the behavior for now.

Bug: T313840
Change-Id: Iba344087b95172da54940e83e89ef2b6e2e36b5d