api: codexicons: make names parameter an enum of all available values
Improves discovery of icon names as they now show up in ApiSandbox.
Avoids having to consult the code directly to identify available icons.

Changed 3rd param of CodexModule::getIcons to default to null which
gives all icons, instead of defaulting to [], which gives no icons. Not
really a breaking change since it's not being invoked anywhere with 3rd
param unspecified as that currently just returns an empty result.

Bug: T311099
Change-Id: Idf1b9a5e8715ef87220e123bb9a061ac3a71212a