Refactor UploadBase to put file verification in a separate class
Hopefully refactoring will make it easier to make changes to this
class in the future.

This only covers verifying the contents of the actual file. Tests
related to on wiki conditions like permissions are still the responsibility
of UploadBase.

Note: This now requires that all subclasses of UploadBase that override
the constructor to call parent::__construct().

Change-Id: I3ec1c0a0850609bd2db4b442228c55a5c0131aaa