session: Avoid references to deprecated SessionManager::singleton
Follows-up I060f773a3c (dc5259885e), which left two calls to
SessionManager::singleton which is now a deprecated wrapper that
simply calls $services->getSessionManager.

Use that or the known reference to the same $manager instance directly
instead.

And update an outdated comment in the test for
maintenance/invalidateUserSessions.php.

Bug: T265400
Change-Id: I245860d2a5a9aead21a50dd10f6d15d834c719da