ResourceLoader: Update cssjanus/cssjanus to wikimedia/cssjanus
Packagist automatically indexed all tags from the repo, so
interestingly that means it immediately has both the current version
as well as all past versions available.

(This in contrast to npm or Docker registries, which are their own
primary source for the versions of a given package, and thus
require publishing there explicilty, vs merely notifying it to pull
the repo for updates.)

No rel-note since it is the same version as before.

Ref Iebe1ee249d738c956b8 in php-cssjanus.git.
Ref https://packagist.org/packages/cssjanus/cssjanus
Ref https://packagist.org/packages/wikimedia/cssjanus

Bug: T411046
Depends-On: I5345056a8af729a7b5d15d71dc6ba01b16c4cd49
Change-Id: Ibc378ff282ff710ac6ed8150ccba68a5f164557e