REST: produce proper OpenAPI specs for transform endpoints
Several endpoints with different parameters and responses map
to the TransformHandler class. Currently, TransformHandler attempts
to deal with this in a way that satisfies all endpoints. The result
is that the generated OpenApi spec does not match any of the
endpoints, making the spec and REST Sandbox less useful for them.

Instead, create "shim" handlers for each endpoint that deal only
with spec generation (and associated parameter validation) in a
way specific to each endpoint, while still deferring all logic to
the TransformHandler class.

The OpenAPI spec generated by the shim handlers in this change make
the REST Sandbox functional for these endpoints, but do not include
every optional parameter supported by the endpoints. These can be
added in a followup change.

Bug: T392608
Change-Id: I171456af422a11c68f74d9eeff196f63907c80eb