installer: Use input type from params if provided
The type provided to Html::input overrides the one set via
the params array. Therefore, we should pass the type
set via the params to the Html::input function, if it
is set.
This fixes the type of all fields built using getTextBox
being 'text'. This is especially important for password
boxes created via getPasswordBox.

Bug: T394796
Change-Id: Ib68782eacec0cb1993d983537deacea9c901463c