Add lcobucci/jwt to Composer requirements
The (somewhat outdated) 4.1.5 version is chosen for
compatibility with the OAuth extension (which is stuck with
the older version due to T261462). We could use a more
permissive constraint here (such as '^4.0 || ^5.0') but
all other core libraries are pinned to an exact version, and
there isn't that much difference between 4.1.5 and the
current 5.5.0.

No vendor patch needed as it's already there as a dependency of
the OAuth extension.

Bug: T399243
Change-Id: I386bf7dbcc2e01a7bccf5ff87d6aa6a19bad1975