BrokenRedirectConstraint: fail if target changed but redirect is still broken
When the BrokenRedirectConstraint fails, we save the target of the
broken redirect to a hidden variable called
`wpAllowedBrokenRedirectTarget`.
On the next save attempt, the value of this variable will then be
passed to the BrokenRedirectConstraint.

If the redirect target in the new content is the same as the one
that caused the constraint to fail in the last save attempt, the
constraint will pass this time and allow saving the page.
If the redirect target has changed since the last save attempt, the
constraint will be checked again and fail if the new target still
does not exist.

If the current content of the page contains a broken redirect and the
new one the user is trying to save does too, but the new non-existent
target is a different one, the constraint will now also fail. Before
this, it was possible to point an already broken redirect to a
different non-existent page without a warning.

Also updated the search api test so it creates the target before
editing a redirect.

Bug: T395767
Change-Id: Idbea125054a1fe3cd163776d42e60b38f01598ba