jwt: Add wrapper class for using lcobucci/jwt
Adds the JwtCodec service, which can be used to generate, parse
and (to the extent it's possible without any knowledge of the
context) validate JWTs using a centralized key configuration.

The configuration itself is MVP-level and should be improved
later (to support different encryptions, and file paths
instead of key strings).

Bug: T399243
Change-Id: Ib041341940f63c2a96ff37d33008f8883982fe76