RedirectConstraint: warn for redirects pointing to an invalid redirect target
Shows a warning when trying to create a redirect that points to a
page that is an invalid redirect target (currently Special:UserLogout
and any special page in $wgInvalidRedirectTargets).

Also adds an $allowInvalidTarget parameter to
WikitextContentHandler::extractRedirectTargetAndText, so it's
possible to get the redirect target even if it is invalid.

Bug: T384578
Change-Id: I723ce5f94210535e036bcd0234dba36cbbd4a91e