Untangle HTMLComboboxField class from XmlSelect (re-submit)
This allows us to get rid of a single-use feature. The ability to
change the tag name was added in 2015 via I954d3d2 but never used
anywhere else.

Hard-deprecate the method right away because it is not used
anywhere, as far as we can know (according to CodeSearch).

This is a fixed re-submission of I29e67a1 with the mistake fixed.

Bug: T118119
Change-Id: Ib37190efd58e4fd716fda739ac41e7f8bb5a4d94