RecentChange: Provide a list of internal source
Currently in quite a few places code checks against RC_EXTERNAL. In
order to replace these usages with an rc_source equivalent, we should
provide a list of all internal source in order to avoid using an array
of all four sources everytime.

This also makes it easier to potentially add a new core source at some
point in the future if that should be desired.

Bug: T74157
Change-Id: Ie12808bc9cd2e5540d593bcadef041400c9aa3d5