Add "subpageof" parameter to changes list special pages
Add the URL parameter "subpageof" to ChangesListSpecialPage, and thus to
RecentChanges, RecentChangesLinked and Watchlist.

The parameter name is by analogy with CirrusSearch which has a filter
with that name. This is a prefix search with an implicit trailing
slash. I think it better reflects the requested feature and the user's
typical intention than a "prefix" parameter which filters without the
trailing slash.

Since the namespace and title are in the table, the performance should
be better than the join conditions in RCL and Watchlist which are
sometimes used as a substitute for this feature.

Add an explanation for why I am not implementing this in
parseParameters().

The UI can be done separately.

Bug: T23674
Bug: T401743
Change-Id: I083bd581d5656fc659f1d4064b447d25c848b027