Replace @phan-side-effect-free with #[\NoDiscard]
Requires Phan 5.5.2 (https://github.com/phan/phan/releases/tag/5.5.2)
to avoid losing static analysis capabilities.

The #[\NoDiscard] attribute is new in PHP 8.5. For PHP < 8.5 we
already have a no-op stub provided by the symfony/polyfill-php85
package (added in a2a0cb09c4c7ce51f98a).

In Phan, the annotation and the attribute have the same effect
(requiring the return value of the function to be used - e.g. assigned
to a variable or passed as a parameter), but the attribute is named
better (as many of the functions where we want to use it are not in
fact "side-effect-free") and it can be enforced at runtime by PHP 8.5.

Change-Id: I75468f5c2315df0c5865c5ed7ec29861940ab2e5