PHPSessionHandler: Better handle objects stored in the session
The context:

It is possible to store PHP objects in the session data, both through
the PHP $_SESSION and the MediaWiki Session::set(). This is used in
AuthManager::autoCreateUser(), which stores StatusValue objects under
the 'AuthManager::AutoCreateBlacklist' key.

PHPSessionHandler keeps PHP and MediaWiki sessions in sync by
detecting changes in $_SESSION, which it does by comparing the value
when a write is requested to the value that was read from the session
storage at the start of the request (or that was last written).

The problem:

If an object is stored, and PHP session handling is enabled, this
detection did not work correctly, because SessionHandlerInterface
methods all deal with serialized data, and must deserialize it first
to compare the new and old values. The comparison was done with `===`,
and a serialized-and-deserialized object is not the same object any
more.

As a result, PHPSessionHandler would always think that the session
data was changed, and would always cause a session write, on every
request where the session data contained any objects. If warn mode
was enabled, this would also generate spurious warnings about writes
to $_SESSION, which we've seen in T400668.

The fix:

If any objects are involved, compare their serialized representation
instead (using the same serializer as the session storage).

Bug: T402602
Change-Id: I9d67e5b8b617af5138c45d474a323e47c36b626b
(cherry picked from commit 049d750efb08806b6cb7b9e6b7326e2ad63f7d6c)