session: Remember newly created secrets to fix Token::wasNew()
Authentication code relies on checking Token::wasNew() to guess
whether a token mismatch during form submission is caused by a CSRF
attempt, or by the user agent not supporting cookies. This did not
work reliably when anything else accessed the token, since it would
be no longer marked as 'new'.

As a result of this change, attempting to log in with cookies disabled
(with default MediaWiki settings) produces this correct error message:
> Testwiki uses cookies to log in users. You have cookies disabled.
> Please enable them and try again.
instead of this incorrect one:
> There seems to be a problem with your login session; this action
> has been canceled as a precaution against session hijacking.
> Please resubmit the form.

Change-Id: I0eff569051258513b0775583e3d0df3018c844e8