recentchanges: Stop doing a join when purging RecentChanges rows
Purging RecentChanges is very simple and does not need all the fields
and joins required by RecentChange::getQueryInfo(). The joins are
harmful when referential integrity is occasionally broken. However, the
hook RecentChangesPurgeRows may depend on these fields being present.

So, deprecate the hook, and use a different query if there are no
registered handlers of it. Instead add a hook which can add fields to
the query as required.

Bug: T403002
Change-Id: Idc3125616e86580374b4122300c746c49fbe517a