recentchanges: Require callers to opt in to the straight join
The straight join on the actor table is a fairly narrow hack to work
around a MariaDB optimiser issue, applicable only to certain kinds of
queries. When Special:NewPages is requested with a username filter, the
actor table usually needs to be first. AbuseFilter appears to also be
negatively affected by the straight join.

Naive callers with unknown conditions are best served by relying on the
optimiser, RecentChange::getQueryInfo() can't know what table order they
need.

So, require callers to pass STRAIGHT_JOIN_ACTOR to
RecentChange::getQueryInfo() to opt in to the straight join.

Bug: T403222
Bug: T311360
Change-Id: Id6b450ce85ca6fb7acf6de9255051dbb2bfcdb94