recentchanges: Allow registration of known rc_source values
This will be used to construct exclusive queries which filter out
certain rc_source values using rc_source IN (...).

Also allow extensions to register whether an rc_source value indicates
"primary" events. For Flow this would be true, and for Wikibase this
would be false. I think this adequately covers the Special:ActiveUsers
use case and the CheckUser use case, both of which want to exclude
Wikibase but include Flow.

Also add the "revision" option, to replace the
SpecialWatchlistGetNonRevisionTypes hook which is used for the
hidelastrevision filter in order to determine whether rc_this_oldid is
significant.

Bug: T74157
Bug: T403142
Change-Id: I525cfb05f76136c44b60a76f8ed21d65461a8cf4