Unified module metrics for MediaWiki REST API
This work kickstarts the system of unified metrics measurement
between REST and ACTION APIs with unified labels, to enable
better cuts across the data for grafana and application-layer
monitoring and alerts.

This patch only implements the REST API metrics in a way that
enables conceptual unification of the labels with the Action API.
The implementation for the metrics fo rthe Action API will be
done as a followup.

The metrics collect standardized details for both REST and ACTION
APIs. It adds to the already existing labels `path`, `status` and
`method` consistently to both, and sets up the `api_type` (REST_API
or ACTION_API) as a label for easy cross-cutting information.

A conceptual adjustment in this patch is the conept of "module".
This concept is different between REST (where it is considered
a collection of endpoints) and ACTION (where it represents an
endpoint). In order to unify the concepts for the sake of monitoring
the "module" and "endpoint" will represent the same conceptual
ideas: endpoint is the handler, and module is a collection of
endpoints. REST API already has those concepts, and the monitoring
can collect those from the spec. ACTION API required some
representation: action module will be `api_endpoint` and the
`api_module` (representing a 'collection') will for the moment
record the API module's namespace to represent where in the
code it originates from to give us some grouping.

The code also allows for overriding those by sending a specified
module name, which can be done at a later stage.

Bug: T402074
Change-Id: I41b4559c50dd2123ad7a7bc61f3b0bd449f4e252