ApiParse: Add `parser=parsoid|legacy|default` parameter
The existing `parsoid` parameter to ApiParse doesn't allow `parsoid`
to be set to false, so once Parsoid becomes the default parser there
would be no way to use the `parsoid` parameter to disable parsoid and
get a legacy parse.  As such we've deprecated the `parsoid` parameter.

Add a new `parser=parsoid|legacy|default` parameter which is
future-proof in the sense that it will still allow forcing a legacy
parse in the future even once Parsoid is the default.

Bug: T399897
Change-Id: I1de65dfcedac4d05e1c06b399bb7bb15207775ad