Extract Exif from PNG files
There are at least 4 competing ways to embed Exif data into a PNG
file. However the most recent edition of the PNG spec made one
of these the official way. This adds support for that method (and
only that method).

This works very similar to how webp exif extraction works. The Exif
data is essentially a tiff file embedded inside a PNG file, so we
extract that file and then read it as if it is a tiff file.

Bug: T269677
Change-Id: Ie87587a79f71ce98935ea5394db0f450e6856b82