recentchanges: Simplify current-revision filters
The filters which exclude current or non-current revisions allow RC
rows that do not have an associated revision. This was done by rc_type.
Instead, introduce a convention where rc_this_oldid=0 indicates that a
recentchanges row does not have an associated revision, and use this
fact for filtering instead of rc_type.

This simplifies the expression sent to the database and will allow the
hook to be deprecated.

By inspection of the WMF production DBs, I see that rc_this_oldid was
already zero if and only if rc_type was one of the two relevant values
(log or flow), so no change is needed to make this be the case.

Bug: T403142
Change-Id: I1426c632f6ccebfa4dc9386e374006cb885f382e