recentchanges: Add ChangesListFilterGroupContainer
* Add a container class which wraps an array of filter groups.
* Add typed accessors for some groups known to core. These will be
  deleted in the subsequent commit.
* Make the container class be responsible for most kinds of iteration
  over groups. If the group interface knows about a concept, then the
  container can know about it too.
* ChangesListSpecialPage::getStructuredFilterJsData() changed from
  public to protected per class todo comment. No known extension
  callers.
* Fix fixme in ChangesListSpecialPageTest.

The signature of FetchChangesListHook changes, but not enough to break
any extension in code search. I could convert it back to an array just
for the hook, but the only thing using that parameter was SkinMinerva,
which just passes the parameter through to the OldChangesList
constructor, which would just need to put it back in an object.

Bug: T404416
Change-Id: I7e3c3049b55f6fa43c21c48ec105005af85f8649