Implement unified metrics for the Action API
This is a followup for the unified metrics for the API ecosystem.
The previous patch implemented this unified system for the REST API
and this patch implements the same conceptual monitoring for the
Action API, utilizing the conceptual meaning of "module" (collection
of endpoints) and "endpoint" (handler or action api module).

Bug: T402074
Change-Id: Ib9a425db4f4a110e42c09761e6f9c75f6c80440f