SpecialSearchResultsHook should run before calculating totals
This hook can be used to modify the search result set. This hook was
running after some derived data of the search result set had already
been used in the Special:Search page and caused the match numbers to
be out of sync with the results.

By running the hook earlier, we prevent such problems.
We still run after SearchResultsAugment, to preserve existing execution order.
We do move the execution of SpecialSearchCreateLink hook, but there
doesn't seem to be a risk of conflict there.

Bug: T85597
Change-Id: Ie7b51f0a38d9f53433e3aff72432e8fd22c98e4f