ContentHolder: normalize argument names, add default argument values
Added a default fragment name of BODY_FRAGMENT since that is the
most frequent value for get*() and set*() methods.  Even though
the fragment is sometimes the first argument, with PHP named
arguments you can use (eg):

  $contentHolder->setAsHtmlString(html: $html);

to take advantage of the default value for fragment name.

With that in mind, standardized argument names to `$fragment` and
`$html` to facilitate using named arguments when calling these
methods.

Bug: T347062
Change-Id: I59b74b8ce80be29bbe8c4a4b80756fb8c33c5228