Add `leids` parameter to `list=logevents`
Previously, it was not possible to retrieve log entries directly by
their IDs; the only workaround was to scrape `Special:Log`. With this
addition, API clients no longer need to rely on such non-API methods.

Internally, this patch introduces the ability to append a clause like
`AND log_id IN (51,55,59,60)` to the DB query. In fact, this is
significantly faster than executing the query without it.

Bug: T309483
Change-Id: Ia1e2a7a1e7db41074a4f9fc51132a3f39d679693