DeletedContribsPager: Use the UserIdentity object instead of the raw target string
Fixed the query for imported actors and some other potential edge cases.

Unsetting the 'target' field in SpecialDeletedContributions alone should
be sufficient, but I would rather like to keep the behaviour the same
with ContribsPager, which is used by more users and using
`$this->targetUser->getName()` is known to be ok so far.

Also, renamed some variables to match the parent class method signature
to avoid confusion.

Bug: T372444
Bug: T404230
Change-Id: I318ec7f30174087f988536f5196ff81e99241c9b