search: Expose query dependent profiles via api
Search currently has a concept of query-independent profiles,
which influence things like how much article popularity
influences search ranking. Patch adds a second query-dependent
profile intended to influence things like field weights, or
default matching behaviour (and/or, % of terms matches, etc).

This only provides the integration point for offering a list of
profiles to a user and passing that back to the search engine. A
SearchEngine will need to add support for this profile via the
same machinery used for the query-independent profile.

Change-Id: I13bc6788d54ce8f5dc949ebede9618d5cc7252d1