wdio-mediawiki: Wait for mw.loader.using in waitForModuleState()
mw and mw.loader are part of the startup module, but mw.loader.using is
part of the mediawiki.base module, so waiting for mw to exist is not
sufficient to ensure that mw.loader.using() can be called. This should
fix errors like the following:

> Page "before each" hook for Page
> WebDriverError: javascript error: mw.loader.using is not a function

Also fix a typo in a previous timeout message.

Bug: T297346
Change-Id: I0adf31a7d7a0aee575c639ee11d1e39a0a52b1a9