session: Add a mechanism for forcing a refresh
When there is valid authenticating information when loading the
session, but the session store data is missing, or session token
information is missing in such a way that the session store data
can't even be identified (e.g. there's a user token cookie but no
session ID cookie), SessionManager will end up (re)persisting the
session and thus updating the session store data and cookies.

There is, however, no way to do that when the session tokens are
deficient in some other way, but the session store data is
identifiable and correct. Most of the mechanism that can be used
to coustomize session loading (SessionProvider::mergeMetadata(),
SessionProvider::refreshSessionInfo(), SessionCheckInfo) can only
choose between loading the session or rejecting it (which results
in session cookies getting deleted). To support JWTs with expiry
much shorter than the session store expiry, we want the provider
to be able to signal "this is a valid session but needs to be
refreshed".

That needs either some kind of expiry information in SessionInfo
or an explicit refresh flag. Going with the flag here as it seems
simpler.

Also contains some small unrelated SessionManager test fixes.

Bug: T399200
Change-Id: I81de5a77327524dce991275d70b0b3171b658e4e
(cherry picked from commit 3fc25a2fc0abd2cd68d5eb3f83f0926b5921b2bd)