maintenance: use $wgExtensionDirectory in SchemaGenerator
When extensions are located outside of the MediaWiki installation path,
tests descending `AbstractSchemaTestBase` would fail.

When generating the database schema files, the generator emits the
absolute path to the SQL files when it should instead yields a path that
is relative to the root of the extension.

The reason is the path canonicalization does not take in account
extensions can be located outside of IP. In such a case, it erroneously
emits an absolute path. The paths mismatch and the schema test fails.

I have my repositories cloned using the same hierarchy as Gerrit:

  ./
  └── mediawiki/
      ├── core/
      │   └── LocalSettings.php
      └── extensions/
          └── AbuseFilter/

And have set `$wgExtensionDirectory = "../extensions";`.

Change the logic to handle extensions and core independently determined
as whether the json path starts with $wgExtensionDirectory.

For extensions:
* strip $wgExtensionDirectory
* drop the extension name using a split instead of preg_replace.
For core:
* strip $IP

Add a test to cover for extensions being outside of $IP.

With AbuseFilter installed, I run the tests for extensions and core
schemas using:

  ./vendor/bin/phpunit --testsuite=extensions \
    --filter testSchemaChangesHaveAutoGeneratedFiles

Change-Id: Ifc739fc8d5321a5274b72260969a85c83c7d79ed