rdbms: Add tests to SchemaGenerator path normalization
The generated schemas have a comment at the top referring to the path of
the json schema used to generate it. The path is relative to MediaWiki
core and for extensions is relative to the extension root.

Extract the normalization path to a standalone method and add tests
veryfing the behavior for core and an extension.

Change-Id: Ibc0d001143f386c81160a77b29d385f9f035670c