objectcache: Add a hit/miss flag to CachedBagOStuff
The caller might want to know whether a CachedBagOStuff::get() call
was served from the in-process cache or the inner BagOStuff object.
Since the signature of get() is inherited and there's no way to add
more information to it, add a separate method that can called after
get() to tell whether it was a hit or miss.

There is no easy way to support getMulti() (which might return some
values from the in-process cache and others from the inner store),
but it's also not used much, so we don't try to support it (but
throw an exception to avoid someone mistakenly believing that the
method works).

Change-Id: I5188c2c2d96f7d8b489145894aeb8d22a253d9b6