Do not access bundle on non-Parsoid content
It can happen, in the case of non-wikitext ContentModel, that a
ContentHolder is not a Parsoid-ContentHolder despite the ParserOptions
asking for it. Trying to access the ContentHolder PageBundle in these
circumstances trips an assert.
This patch limits the access of the ContentHolder PageBundle to
ContentHolder that are explicitly identified as Parsoid-typed in the
case identified in T404902.

Bug: T404902
Change-Id: I9ae03be4d4147e28d328d22b4e4a74944be2abd3
(cherry picked from commit 752c7513a4f7edabb6ef4b2464347251e5770681)