SECURITY: Disallow 'style' attribute in client-side messages (jqueryMsg)
CVE-2025-67481

The 'style' attribute is difficult to sanitize, and the logic needs to
be updated constantly as new CSS features are developed by browser
vendors. Use cases for the 'style' attribute in client-side messages
are extremely rare, and the maintenance burden is high, so let's just
disallow it. If needed, 'class' and 'id' are still allowed, so these
elements can be targeted by normal stylesheets.

Bug: T251032
Change-Id: I192d238bc961f67427595499144d56bcc1747211