Improve css checking in SVG filter.
Use more in depth CSS parsing to better check for CSS rules
that cause external resources to be loaded.

Note: This removes some checks for ancient issues. CSS expression
was dropped in MSIE 8 (2009). The -o-link stuff was no longer a
vulnerability as of opera 11.01 (2011) and dropped entirely in
Opera 15 (2013). These browsers do not even support TLS 1.2.

This also adds checks for the cursor attribute which was missed
previously and supports specifying a url.

Bug: T85085
Change-Id: I526a815f8ec8e357abae3dcf5ef4e6c3081ab7c0