Login page: make password reset discoverable for blocked users
This used to be hidden 'when you cannot use it'. But this can be
confusing, as you cannot find out WHY the link is missing. Users that
are IP blocked might be unable to discover they cannot reset their
password because they are blocked.

Since recently the check was split into isEnabled and isAllowed, so
use isEnabled for presence of the link, and isAllowed can be used by
the password reset process to check if it is allowed and present you
with the error messages.

Bug: T405167
Change-Id: I58b1460b225040e95bdae6e74af961e508ae778b