Add support for displaying DigitalSourceType image metadata
This metadata field is used for recording the digital source of
an image, inclduing if it was created by generative AI.

In order to actually extract it we also need a change in XMPReader
I8ff4b4170. Long term we will hopefully add support for C2PA which
also uses the same metadata schema for this but is more common.

Bug: T405138
Change-Id: I3a25669ba83f97c2ccf2bef156d60cd214fa646c