Deprecate User::getInstanceForUpdate()
It led to an overuse of exclusive locks, often by callers which
don't need locking at all. The name was a reference to the SQL
option FOR UPDATE, but this wasn't obvious and it was easy to
assume that you are just getting a primary instance. Provide an
alternative with a clearer name.

Updating callers is left to follow-ups, as changing locking
behavior often leads to unexpected consequences, so it's better
done in small separate patches.

Bug: T405231
Change-Id: Iffea7ffdf5cc25de49050e6f48d80417cc23fbae
(cherry picked from commit ab50bb3847ed28ca9b7b29797519745dbe0a0863)