session: Fix date handling for JWT cookies
The code assumed lcobucci/jwt returns a scalar array that matches
the raw JSON content of the JWT, but it actually converts date
claims to DateTimeImmutable. That is nice when working with the
library directly, but since we don't want to depend on
library-specific behavior, it's better to undo it for JwtCodec.

Follows-Up: Ib041341940f63c2a96ff37d33008f8883982fe76
Bug: T399243
Bug: T399200
Change-Id: I10d9d77e15edc0712f90772d6fedab99ff9dbd36