Do not lock user table in options APIs
ApiOptionsBase (used by the action=options API and the
two GlobalPreferences APIs) obtained an exclusive lock
on the user record before changing options. This wasn't
really useful (we are changing user options, which are in
a different table) and at times became a performance
bottleneck.

The patch replaces getUserForUpdate[OrNull] with
getUserFromPrimary[OrNull] (which don't lock). We probably
don't even need a primary instance (the user object is
with UserOptionsManager or GlobalPreferencesFactory, which
only care about the user ID) but maybe useful in edge cases
with a freshly autocreated user, so leaving that for now.

Also change ApiOptions::commitChanges(), which was the one
exception that actually saved the user (and so needed an
up-to-date user record) but there isn't any obvious reason
to do that - it's probably just a leftover from
User::saveSettings() being the way to save user options in
the past.

The old methods are retained as aliases to avoid a circular
dependency on Ia0db0c482e1ecf61d192948e5d35 and will be
removed in a follow-up.

Part of the ApiOptionsBase changes for this patch accidentally
ended up in Iffea7ffdf5cc25de49050e6f48d80417cc23fbae.

Bug: T405231
Bug: T405112
Change-Id: Ic634e2d153a623f4db16357623bb0e664da25cca